Move app/models.py separate user.py model file, change import of models from models.py to app/models folder
