+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+from .user import User
+from .word import Word
+from .drawing import Drawing
+from .guess import Guess
